Refactor the directory hierarchy
The existing hierarchy had a bunch of directories with relatively few
files in them that were grouped by a relatively obscure principle.
This moves most files to the lib/src directory.

This also uses the built-in types for SourceSpan and SourceLocation
rather than defining custom types.
