+++ .gitignore
@@ -5,7 +5,7 @@ lib/src/vendor
 node_modules
 npm-debug.log*
 package-lock.json
-spec/sandbox
+test/sandbox
 
 # Editors
 .idea
