+++ lib/src/async-compiler.ts
