+++ lib/src/exception.ts
@@ -2,7 +2,9 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-import {SourceSpan} from './span';
+import * as proto from './vendor/embedded-protocol/embedded_sass_pb';
+import {SourceSpan} from './vendor/sass';
+import {deprotofySourceSpan} from './deprotofy-span';
 
 /**
  * An exception thrown by Sass.
@@ -36,4 +38,21 @@ export class SassException extends Error {
   toString() {
     return this.formatted;
   }
+}
+
+/**
+ * Creates a SassException from the given protocol `buffer`. Throws if the
+ * buffer has invalid fields.
+ */
+export function deprotofyException(
+  buffer: proto.OutboundMessage.CompileResponse.CompileFailure
+): SassException {
+  const span = buffer.getSpan();
+
+  return new SassException(
+    buffer.getMessage(),
+    buffer.getFormatted(),
+    span ? deprotofySourceSpan(span) : undefined,
+    buffer.getStackTrace()
+  );
 }
