+++ lib/src/packet-transformer.ts
