+++ lib/src/value/index.ts
